Create MixerAgent.java
An auxiliary host named as Mixers is developed to be part of the anonymiser system. 
The owner sends its agent to the anonymiser system that provides a transient virtual owner for the agent in each step of its itinerary. 
This means that any host in the network cannot learn the true identity of the agent owner and the path which the agent has traversed through so far. 

Please refer to the following journal article published in IET Information Security Journal for more information:
F. Raji, B. T. Ladani, Anonymity and Security for Autonomous Mobile Agents, IET Information Security, Special Issue on Multi-Agent and Distributed Information Security, Vol.  4, No. 4, pp. 397 - 410, 2010.